backend/requirements.txt
# FastAPI and related dependencies
fastapi>=0.95.0,<0.100.0
uvicorn[standard]>=0.22.0,<0.30.0
pydantic>=1.10.7,<2.0.0
starlette>=0.27.0,<0.30.0

# HTTP clients
httpx>=0.24.0,<0.30.0

# MongoDB driver
motor>=3.1.1,<4.0.0

# Authentication and security
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=4.0.1,<5.0.0
python-multipart>=0.0.5,<0.1.0

# Environment and configuration
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0,<7.0.0

# Logging
coloredlogs>=15.0.1,<16.0.0

# Utilities
email-validator>=2.0.0,<3.0.0
Jinja2>=3.1.2,<4.0.0

# Data processing
aiofiles>=23.1.0,<24.0.0
orjson>=3.8.10,<4.0.0
pydantic-core>=2.10.0,<3.0.0
requests>=2.31.0,<3.0.0

ultravox-client